Update with ADSK developments
- Mutliple minor changes and fixes
- Fix some inconsistencies in some License Headers (only Simplified BSD
  same format in all files)
- Added model attachnemnts
- Added unit tests
- Use CMake as build system now
- Add XSDs for core- and other extension-specification and xmllint as
  schema-checking tool

+++ Include/Common/NMR_Assertion.h
@@ -37,7 +37,7 @@ Native Model Repair algorithms and modules.
 #ifdef NMR_USEASSERTIONS
 #ifdef NMR_UNITTESTS
 #include "CppUnitTest.h"
-#define __NMRASSERT(value)  Microsoft::VisualStudio::CppUnitTestFramework::Assert::IsTrue (((int) (value)) != 0);
+#define __NMRASSERT(value)  Microsoft::VisualStudio::CppUnitTestFramework::Assert::IsTrue (((uintptr_t) (value)) != 0);
 
 #else
 #include <cassert>
